Revert "64-bit: more pointer to int fixes"
This reverts commit 1ce0060197d44ef658e675299e5d9c88d9a03724.

+++ runtime/mirror/object.h
@@ -175,7 +175,7 @@ class MANAGED Object {
   void SetFieldObject(MemberOffset field_offset, const Object* new_value, bool is_volatile,
                       bool this_is_valid = true) SHARED_LOCKS_REQUIRED(Locks::mutator_lock_) {
     VerifyObject(new_value);
-    SetField32(field_offset, reinterpret_cast<uintptr_t>(new_value), is_volatile, this_is_valid);
+    SetField32(field_offset, reinterpret_cast<uint32_t>(new_value), is_volatile, this_is_valid);
     if (new_value != NULL) {
       CheckFieldAssignment(field_offset, new_value);
       WriteBarrierField(this, field_offset, new_value);
